requirements.txt

streamlit
plotly
pandas
backtrader
websocket-client
numpy

python-binance
matplotlib

